Rename and reorganise Core folder
Ticket: #180
