+++ src/app/Core/FilterServices/checkIfFilterIsSet.service.ts
