+++ src/app/Core/getCurrentDate.service.ts
@@ -2,7 +2,7 @@
  * @desc Generate the current date and time in a fixed format (YYYY-MM-DDThhmmss)
  * @returns {string} - string with the current date and time
  */
-export function getFormattedDate(): string {
+export function getCurrentDate(): string {
     const today = new Date();
     const y = today.getFullYear();
     const m = today.getMonth() + 1;
