+++ src/app/Core/getTableFromPath.service.ts
@@ -5,7 +5,7 @@ import { FilterType } from "../Shared/Model/Filter.model";
  * @param {string} path - URL path
  * @returns {FilterType[]} - list with the row and column from the URL path
  */
-export function getTableFromPath(path: string): FilterType[] {
+export function getFeaturesFromPath(path: string): FilterType[] {
     const searchParams = new URLSearchParams(path);
     const rowParam: FilterType = searchParams.get("row") as FilterType || "";
     const colParam: FilterType = searchParams.get("column") as FilterType || "";
