+++ src/app/Main/Body/QueryPage/Drawer/Filter/SelectorList-Selector.component.tsx
@@ -1,7 +1,7 @@
 import React, { useContext } from "react";
 import { ValueType } from "react-select";
 import { useTranslation } from "react-i18next";
-import { CheckIfSingleFilterIsSet } from "../../../../../Core/FilterServices/checkIfFilter.service";
+import { CheckIfSingleFilterIsSet } from "../../../../../Core/FilterServices/checkIfFilterIsSet.service";
 import { DataContext } from "../../../../../Shared/Context/DataContext";
 import { FilterContext } from "../../../../../Shared/Context/FilterContext";
 import {
